+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # Braindrop ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Fixed doubling-up of the raindrop details panel's help when the raindrop's
+  tags have focus. ([#39](https://github.com/davep/braindrop/issues/39))
+
 ## v0.2.0
 
 **Released: 2025-01-05**