+++ ChangeLog.md
@@ -10,6 +10,7 @@
   ([#43](https://github.com/davep/braindrop/pull/43))
 - Added <kbd>v</kbd> as a global key for visiting a currently-highlighted
   link. ([#43](https://github.com/davep/braindrop/pull/43))
+- Added public/private icons to the raindrops list.
 
 ## v0.2.0