+++ ChangeLog.md
@@ -11,6 +11,8 @@
 - Added <kbd>v</kbd> as a global key for visiting a currently-highlighted
   link. ([#43](https://github.com/davep/braindrop/pull/43))
 - Added public/private icons to the raindrops list.
+- Added public/private icons and the collection name to the raindrop detail
+  view.
 
 ## v0.2.0